fix classes issues

+++ classes/VectorOp.h
@@ -1,7 +1,6 @@
 // VectorOp.hpp
 // declearation of struct VectorOp
 
-#include <bits/stdc++.h>
 #include "Segment.h"
 #include "Point.h"
 
@@ -13,10 +12,10 @@ using namespace std;
 class VectorOp // Copper Assembly共用
 {
 public:
-    static float cross(const Point, const Point); // 向量外積
+    static double cross(const Point, const Point); // 向量外積
     static double cross1(const Point, const Point, const Point); //共起點向量外積
-    static float cross_point (const Point, const Point, const Point);
-    static float dot(const Point, const Point);   // 向量積
+    static double cross_point (const Point, const Point, const Point);
+    static double dot(const Point, const Point);   // 向量積
     static Point orth_Cswap(const Point); //垂直順時鐘(向量)
     static Point orth_CCswap(const Point); //垂直逆時鍾(向量)
 };
